Refactor: change the namespace of shifting folder files.
